Fix complete backend

+++ board-hub-backend/src/modules/loan/application/use-cases/delete-loan.use-case.ts
@@ -1,5 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
+
 import { LOAN_REPOSITORY, type ILoanRepository } from "../../domain/interfaces/loan.repository.interface";
+import { LoanNotFoundException } from '../../domain/exceptions/loan.exceptions';
 
 @Injectable()
 export class DeleteLoanUseCase {
@@ -9,13 +11,14 @@ export class DeleteLoanUseCase {
     ) { }
 
     async execute(id: string): Promise<boolean> {
-        
-        try {
-            await this.loanRepository.delete(id);
-            return true;
-        } catch (error) {
-            
-            return false;
+
+        const loan = await this.loanRepository.findById(id)
+
+        if(!loan || loan.isDeleted){
+            throw new LoanNotFoundException(id);
         }
+        
+        await this.loanRepository.delete(id);
+        return true;
     }
 }
